// hrm/src/app/Model/schedaAnagraficaCompleta/scheda-anagrafica-completa.ts
export class SchedaAnagraficaCompleta {
  name: string;
  lastname: string;
  age: number;
  email: string;
  province: string;
  genere: string;
  annoDiNascita: number;
  common: string;
  sigla_nazionale: string;
  phone: number;
  skype: string;
  training: string;
  ambito: string;
  cv_disponibile: boolean;
  fonte_cv: string;
  skills_txt: string;
  tipo_candidatura: string;
  status: string;
  protectedCategory: boolean;
  informativaGDPR: string;
}